Manipulating 100% dataframe
The notebook was started a couple of days ago, but the first commit is in 03/11. 
In the first step, the sentiment extraction was through the sparkNLP library. In this stage numerous error were faced (95% out of memory, 5% serialization problems - udf's)
Models for embeddings such as Universal Enconder, Bert, Roberta were not able to be performed for memory issues (memory  error when creating the pipeline), other models like vader using udf also resulted in memory issue and serialization issue. 

The model chosen for the sentiment extraction so far is using the Glove embeddings and a pretrained sentiment model from sparknlp, however the train set is not based on tweets but reviews. The library did not had any pretrained model for tweets (Using glove embeddings). 
Pretrained models that have tweets as  base required more sofisticated embedding which were attempted but resulted in memory issues.